+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FriendDbStorage.java
@@ -18,25 +18,25 @@ public class FriendDbStorage implements FriendDao {
     }
 
     @Override
-    public void saveFriend(int userId, int friendId) {
+    public void saveFriend(Long userId, Long friendId) {
         String sql = "INSERT INTO friends (user_id, friend_id) VALUES (?, ?);";
         jdbcTemplate.update(sql, userId, friendId);
     }
 
     @Override
-    public void deleteFriend(int userId, int friendId) {
+    public void deleteFriend(Long userId, Long friendId) {
         String sql = "DELETE FROM friends WHERE user_id = ? AND friend_id = ?;";
         jdbcTemplate.update(sql, userId, friendId);
     }
 
     @Override
-    public List<Integer> getFriendsByUserId(int userId) {
+    public List<Long> getFriendsByUserId(Long userId) {
         String sql = "SELECT friend_id FROM friends WHERE user_id = ?;";
-        return jdbcTemplate.queryForList(sql, Integer.class, userId);
+        return jdbcTemplate.queryForList(sql, Long.class, userId);
     }
 
     @Override
-    public boolean containsInStorage(int filmId, int friendId) {
+    public boolean containsInStorage(Long filmId, Long friendId) {
         String sqlQuery = "SELECT count(*) FROM friends WHERE user_id = ? AND friend_id;";
         int result = jdbcTemplate.queryForObject(sqlQuery, Integer.class, filmId, friendId);
         return result == 1;
